Исправил все, как было описано в PR
