+++ dotNet module/Task_2/IRemind.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Task_2
+{
+    /// <summary>
+    /// интерфейс напоминания
+    /// </summary>
+    public interface IRemind
+    {
+        public DateTime EventTime { get; set; }
+    }
+}
